requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0

# Jupyter and Notebook support
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# HTTP and API
requests>=2.28.0
httpx>=0.24.0

# Environment and configuration
python-dotenv>=0.19.0

# Data processing and analysis
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Text processing
tiktoken>=0.4.0
pypdf>=3.0.0
python-docx>=0.8.11
docx2txt>=0.8

# Vector stores and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Utilities
tqdm>=4.64.0
pyyaml>=6.0